+++ website-rota/del_tabelas.php
@@ -1,10 +1,16 @@
 <?php
 session_start();
+
+ if(!isset($_SESSION['id'])){
+      header('location: area_cliente.php');
+      exit;
+ }
+ 
 include_once 'dao/clienteDAO.class.php';
 $cliDAO = new ClienteDAO();
 $cliDAO->deletarTabelas();
 
-//código qie deleta oque a dentro da pasta arquivos/
+//código que deleta oque a dentro da pasta arquivos/
 $caminho = "arquivos/";
 system("rm -rf $caminho");
